Adds Stopwatch class for measuring thread execution time

+++ src/physical-layer-core/receive-multicore-worker/receive-multicore-worker.factory.js
@@ -6,11 +6,13 @@
         .registerFactory('PhysicalLayerCore.ReceiveMulticoreWorker', _ReceiveMulticoreWorker);
 
     _ReceiveMulticoreWorker.$inject = [
-        'PhysicalLayerCore.ReceiveMulticoreWorkerThread'
+        'PhysicalLayerCore.ReceiveMulticoreWorkerThread',
+        'Common.StopwatchBuilder'
     ];
 
     function _ReceiveMulticoreWorker(
-        ReceiveMulticoreWorkerThread
+        ReceiveMulticoreWorkerThread,
+        StopwatchBuilder
     ) {
         var ReceiveMulticoreWorker;
 
@@ -19,6 +21,7 @@
                 throw ReceiveMulticoreWorker.MULTICORE_SUPPORT_IS_NOT_ENABLED_EXCEPTION;
             }
 
+            this.$$stopwatch = StopwatchBuilder.build();
             var js = ReceiveMulticoreWorkerThread.getJavaScriptCode();
             var blob = new Blob(
                 [ js ],
@@ -27,17 +30,24 @@
 
             this.$$worker = new Worker(URL.createObjectURL(blob));
             this.$$worker.onmessage = this.onMessage.bind(this);
-
-            console.log('before post');
-            this.$$worker.postMessage("Hello, I just send you message");
-            console.log('after post');
         };
 
         ReceiveMulticoreWorker.MULTICORE_SUPPORT_IS_NOT_ENABLED_EXCEPTION = 'Multicore support is not enabled';
 
         ReceiveMulticoreWorker.prototype.onMessage = function(oEvent) {
-            console.log("Worker finished: " + oEvent.data);
-            this.$$worker.terminate();
+            var m = oEvent.data;
+
+            if (m === 'ready') {
+                console.log(':: thread started ::');
+                this.$$stopwatch.start();
+                this.$$worker.postMessage("some crazy sine sum");
+            } else {
+                console.log(":: thread stopped ::");
+                console.log("Thread message: " + oEvent.data);
+                console.log("Thread time: " + this.$$stopwatch.stop().getDuration(true) + ' seconds');
+                this.$$worker.terminate();
+            }
+
         };
 
         return ReceiveMulticoreWorker;
